refactor: Auth files
